# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Simple Light
  description: "Simple application (OTP) for API course at University of Szeged"
  contact: { 
    "name": "Berki Tamás",
    "email": "Berki.Tamas.Jozsef@stud.u-szeged.hu",
    "url": "http://btj.hu"
  }
  license: {
    "name": "MIT",
    "url": "https://github.com/berkitamas/simplelight/blob/master/LICENSE"
  }
host: localhost:10010
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-Session-ID
security:
  - ApiKeyAuth: []
paths:
...
  /users/{userId}:
    x-swagger-router-controller: users
    parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        type: string
    get:
      tags:
        - user
      summary: Retrieve a user
      description: Retrieves a user with a given ID
      operationId: retrieveUser
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/User"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
          
  # Card management #
...
  /users/{userId}/parkingTickets:
    x-swagger-router-controller: parkings
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      tags:
        - parking
        - user
      summary: Retrieve all parking tickets
      description: Returns all parking tickets of the given user
      operationId: retrieveAllParkingTickets
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/ParkingTicket"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - parking
        - user
      summary: Add parking ticket
      description: Creates a new parking ticket for the given user 
      operationId: addParkingTicket
      parameters:
        - name: ticket
          in: body
          required: true
          description: Details of the parking ticket
          schema:
            $ref: "#/definitions/ParkingTicket"
      responses:
        201:
          description: Successfully created
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{userId}/parkingTickets/{ticketId}:
    x-swagger-router-controller: parkings
    parameters:
      - name: ticketId
        in: path
        required: true
        description: ID of the ticket
        type: string
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      tags:
        - parking
        - user
      summary: Retrieve parking ticket
      description: Returns a ticket from the given ids
      operationId: retrieveParkingTicket
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ParkingTicket"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or ticket with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{userId}/parkingTickets/{ticketId}/stop:
    x-swagger-router-controller: parkings
    parameters:
      - name: ticketId
        in: path
        required: true
        description: ID of the ticket
        type: string
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    post:
      tags:
        - parking
        - user
      summary: Stop parking ticket
      description: Stops the given parking ticket of the user and bills the amount
      operationId: stopParkingTicket
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ParkingTicket"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or ticket with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /parkings/{cityId}:
    x-swagger-router-controller: parkings
    parameters:
      - name: cityId
        in: path
        required: true
        description: ID of the city
        type: string
    get:
      tags:
        - parking
      security: []
      summary: Retrieve all parking spot in city
      description: Returns all parking spot in the city
      operationId: retrieveAllParkingSpotInCity
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/ParkingSpot"
        404:
          description: City not found
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  User:
    properties:
      _id:
        type: string
      username:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      motorwayTickets:
        type: array
        items:
          $ref: "#/definitions/MotorwayTicket"
      parkingTickets:
        type: array
        items:
          $ref: "#/definitions/ParkingTicket"
      cinemaTickets:
        type: array
        items:
          $ref: "#/definitions/CinemaTicket"
...
  MotorwayTicket:
    required:
      - cardId
      - licensePlate
      - priceCategory
    properties:
      _id:
        type: string
      userId:
        type: string
      cardId:
        type: string
      licensePlate:
        type: string
      priceCategory:
        type: string
        enum: [Weekly, Monthly, Yearly]
      purchaseDate:
        type: string
        format: date-time
  ParkingTicket:
    required:
      - cardId
      - licensePlate
      - zoneId
    properties:
      _id:
        type: string
      userId:
        type: string
      cardId:
        type: string
      licensePlate:
        type: string
      zoneId:
        type: string
      price:
        type: string
      stopTime:
        type: string
        format: date-time
      purchaseDate:
        type: string
        format: date-time
  CinemaTicket:
    required:
      - cardId
      - showId
      - row
      - seat
    properties:
      _id:
        type: string
      userId:
        type: string
      cardId:
        type: string
      showId:
        type: string
      row:
        type: number
      seat:
        type: number
      purchaseDate:
        type: string
        format: date-time
...
  City:
    properties:
      _id:
        type: string
      name:
        type: string
      posX:
        type: number
      posY:
        type: number
  ParkingSpot:
    properties:
      _id:
        type: string
      city:
        $ref: "#/definitions/City"
      zone:
        type: string
      price:
        type: string
        format: per quarter hour
      areaPoly:
        type: array
        items:
          $ref: "#/definitions/Point"
  Point:
    properties:
      x:
        type: number
      y:
        type: number
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string